+++ client/CMakeLists.txt
@@ -30,7 +30,8 @@ set(client_SRCS
 	gui/CGuiHandler.cpp
 	gui/CIntObject.cpp
 	gui/CursorHandler.cpp
-	gui/InterfaceEventDispatcher.cpp
+	gui/EventDispatcher.cpp
+	gui/EventsReceiver.cpp
 	gui/InterfaceObjectConfigurable.cpp
 	gui/FramerateManager.cpp
 	gui/NotificationHandler.cpp
@@ -164,8 +165,9 @@ set(client_HEADERS
 	gui/CGuiHandler.h
 	gui/CIntObject.h
 	gui/CursorHandler.h
+	gui/EventDispatcher.h
+	gui/EventsReceiver.h
 	gui/InterfaceObjectConfigurable.h
-	gui/InterfaceEventDispatcher.h
 	gui/FramerateManager.h
 	gui/MouseButton.h
 	gui/NotificationHandler.h
